- rename authenticate.js to authenticate.service.js
