+++ server/middleware/authMiddleware.js
@@ -1,4 +1,4 @@
-import authenticate from '@auth/authenticate';
+import authenticate from '@auth/authenticate.service';
 
 // Authentication middleware is group of two middleware
 const middlewareArray = [
